fix(cloud): distinguish invalid volume selection input
